feature: Implement user domain
